+++ BraintreeCore/src/main/java/com/braintreepayments/api/AnalyticsClient.kt
@@ -21,7 +21,6 @@ internal class AnalyticsClient @VisibleForTesting constructor(
     private val workManager: WorkManager,
     private val deviceInspector: DeviceInspector
 ) {
-    private var lastKnownAnalyticsUrl: String? = null
 
     constructor(context: Context) : this(
         BraintreeHttpClient(),
@@ -50,7 +49,6 @@ internal class AnalyticsClient @VisibleForTesting constructor(
         timestamp: Long,
         authorization: Authorization
     ): UUID {
-        lastKnownAnalyticsUrl = configuration.analyticsUrl
         scheduleAnalyticsWrite("android.$eventName", timestamp, authorization)
         return scheduleAnalyticsUpload(configuration, authorization, sessionId, integration)
     }
@@ -124,14 +122,15 @@ internal class AnalyticsClient @VisibleForTesting constructor(
                 val analyticsEventDao = analyticsDatabase.analyticsEventDao()
                 val events = analyticsEventDao.getAllEvents()
                 if (events.isNotEmpty()) {
-                    val metadata = deviceInspector.getDeviceMetadata(context, sessionId, integration)
+                    val metadata = deviceInspector.getDeviceMetadata(context, configuration, sessionId, integration)
                     val analyticsRequest = serializeEvents(authorization, events, metadata)
-                    configuration?.analyticsUrl?.let { analyticsUrl ->
-                        httpClient.post(
-                            analyticsUrl, analyticsRequest.toString(), configuration, authorization
-                        )
-                        analyticsEventDao.deleteEvents(events)
-                    }
+                    httpClient.post(
+                        FPTI_ANALYTICS_URL,
+                        analyticsRequest.toString(),
+                        configuration,
+                        authorization
+                    )
+                    analyticsEventDao.deleteEvents(events)
                 }
                 ListenableWorker.Result.success()
             } catch (e: Exception) {
@@ -141,14 +140,19 @@ internal class AnalyticsClient @VisibleForTesting constructor(
     }
 
     fun reportCrash(
-        context: Context?, sessionId: String?, integration: String?, authorization: Authorization?
+        context: Context?,
+        configuration: Configuration?,
+        sessionId: String?,
+        integration: String?,
+        authorization: Authorization?
     ) {
-        reportCrash(context, sessionId, integration, System.currentTimeMillis(), authorization)
+        reportCrash(context, configuration, sessionId, integration, System.currentTimeMillis(), authorization)
     }
 
     @VisibleForTesting
     fun reportCrash(
         context: Context?,
+        configuration: Configuration?,
         sessionId: String?,
         integration: String?,
         timestamp: Long,
@@ -157,57 +161,66 @@ internal class AnalyticsClient @VisibleForTesting constructor(
         if (authorization == null) {
             return
         }
-        val metadata = deviceInspector.getDeviceMetadata(context, sessionId, integration)
+        val metadata = deviceInspector.getDeviceMetadata(context, configuration, sessionId, integration)
         val event = AnalyticsEvent("android.crash", timestamp)
         val events = listOf(event)
         try {
             val analyticsRequest = serializeEvents(authorization, events, metadata)
-            lastKnownAnalyticsUrl?.let { analyticsUrl ->
-                httpClient.post(
-                    analyticsUrl,
-                    analyticsRequest.toString(),
-                    null,
-                    authorization,
-                    HttpNoResponse()
-                )
-            }
+            httpClient.post(
+                FPTI_ANALYTICS_URL,
+                analyticsRequest.toString(),
+                null,
+                authorization,
+                HttpNoResponse()
+            )
         } catch (e: JSONException) { /* ignored */
         }
     }
 
     @Throws(JSONException::class)
     private fun serializeEvents(
-        authorization: Authorization?, events: List<AnalyticsEvent>, metadata: DeviceMetadata
+        authorization: Authorization?,
+        events: List<AnalyticsEvent>, metadata: DeviceMetadata
     ): JSONObject {
-        val requestObject = JSONObject()
+        val batchParamsJSON = metadata.toJSON()
         authorization?.let {
             if (it is ClientToken) {
-                requestObject.put(AUTHORIZATION_FINGERPRINT_KEY, it.bearer)
+                batchParamsJSON.put(AUTHORIZATION_FINGERPRINT_KEY, it.bearer)
             } else {
-                requestObject.put(TOKENIZATION_KEY, it.bearer)
+                batchParamsJSON.put(TOKENIZATION_KEY, it.bearer)
             }
         }
 
-        requestObject.put(META_KEY, metadata.toJSON())
-        val eventObjects = JSONArray()
-        var eventObject: JSONObject
+        val eventsContainerJSON = JSONObject()
+        eventsContainerJSON.put(BATCH_PARAMS_KEY, batchParamsJSON)
+
+        val eventParamsJSON = JSONArray()
         for (analyticsEvent in events) {
-            eventObject = JSONObject()
-                .put(KIND_KEY, analyticsEvent.name)
+            val singleEventJSON = JSONObject()
+                .put(EVENT_NAME_KEY, analyticsEvent.name)
                 .put(TIMESTAMP_KEY, analyticsEvent.timestamp)
-            eventObjects.put(eventObject)
+                .put(TENANT_NAME_KEY, "Braintree")
+            eventParamsJSON.put(singleEventJSON)
         }
-        requestObject.put(ANALYTICS_KEY, eventObjects)
-        return requestObject
+        eventsContainerJSON.put(EVENT_PARAMS_KEY, eventParamsJSON)
+
+        // Single-element "events" array required by FPTI formatting
+        val eventsArray = JSONArray(arrayOf(eventsContainerJSON))
+        return JSONObject().put(EVENTS_CONTAINER_KEY, eventsArray)
     }
 
     companion object {
-        private const val ANALYTICS_KEY = "analytics"
-        private const val KIND_KEY = "kind"
-        private const val TIMESTAMP_KEY = "timestamp"
-        private const val META_KEY = "_meta"
-        private const val TOKENIZATION_KEY = "tokenization_key"
+        private const val FPTI_ANALYTICS_URL = "https://api-m.paypal.com/v1/tracking/batch/events"
+
+        private const val EVENTS_CONTAINER_KEY = "events"
+        private const val BATCH_PARAMS_KEY = "batch_params"
         private const val AUTHORIZATION_FINGERPRINT_KEY = "authorization_fingerprint"
+        private const val TOKENIZATION_KEY = "tokenization_key"
+        private const val EVENT_PARAMS_KEY = "event_params"
+        private const val EVENT_NAME_KEY = "event_name"
+        private const val TIMESTAMP_KEY = "t"
+        private const val TENANT_NAME_KEY = "tenant_name"
+
         private const val INVALID_TIMESTAMP: Long = -1
         const val WORK_NAME_ANALYTICS_UPLOAD = "uploadAnalytics"
         const val WORK_NAME_ANALYTICS_WRITE = "writeAnalyticsToDb"