+++ Demo/src/main/java/com/braintreepayments/demo/CardFragment.java
@@ -17,7 +17,6 @@ import androidx.navigation.fragment.NavHostFragment;
 
 import com.braintreepayments.api.AmericanExpressClient;
 import com.braintreepayments.api.AmericanExpressRewardsBalance;
-import com.braintreepayments.api.BraintreeClient;
 import com.braintreepayments.api.Card;
 import com.braintreepayments.api.CardClient;
 import com.braintreepayments.api.CardNonce;
@@ -71,12 +70,11 @@ public class CardFragment extends BaseFragment implements OnCardFormSubmitListen
     public void onCreate(Bundle onSaveInstanceState) {
         super.onCreate(onSaveInstanceState);
 
-        BraintreeClient braintreeClient = getBraintreeClient();
-        americanExpressClient = new AmericanExpressClient(braintreeClient);
-        cardClient = new CardClient(braintreeClient);
-        threeDSecureClient = new ThreeDSecureClient(braintreeClient);
+        americanExpressClient = new AmericanExpressClient(requireContext(), super.getAuthStringArg());
+        cardClient = new CardClient(requireContext(), super.getAuthStringArg());
+        threeDSecureClient = new ThreeDSecureClient(requireContext(), super.getAuthStringArg());
 
-        dataCollector = new DataCollector(braintreeClient);
+        dataCollector = new DataCollector(requireContext(), super.getAuthStringArg());
 
         if (onSaveInstanceState != null) {
             threeDSecureRequested = onSaveInstanceState.getBoolean(EXTRA_THREE_D_SECURE_REQUESTED);
@@ -126,26 +124,20 @@ public class CardFragment extends BaseFragment implements OnCardFormSubmitListen
 
     private void configureCardForm() {
         final AppCompatActivity activity = (AppCompatActivity) getActivity();
-        BraintreeClient braintreeClient = getBraintreeClient();
 
-        braintreeClient.getConfiguration((configuration, configError) -> {
-            if (configuration != null) {
-                cardForm.cardRequired(true)
-                        .expirationRequired(true)
-                        .cvvRequired(configuration.isCvvChallengePresent())
-                        .postalCodeRequired(configuration.isPostalCodeChallengePresent())
-                        .mobileNumberRequired(false)
-                        .actionLabel(cardFormActionLabel)
-                        .setup(activity);
+        // TODO: Configure card form via settings
+        cardForm.cardRequired(true)
+                .expirationRequired(true)
+                .cvvRequired(true)
+                .postalCodeRequired(true)
+                .mobileNumberRequired(false)
+                .actionLabel(cardFormActionLabel)
+                .setup(activity);
 
-                if (getArguments().getBoolean(MainFragment.EXTRA_COLLECT_DEVICE_DATA, false)) {
-                    dataCollector.collectDeviceData(activity,
-                            (deviceData, e) -> this.deviceData = deviceData);
-                }
-            } else {
-                handleError(configError);
-            }
-        });
+        if (getArguments().getBoolean(MainFragment.EXTRA_COLLECT_DEVICE_DATA, false)) {
+            dataCollector.collectDeviceData(activity,
+                    (deviceData, e) -> this.deviceData = deviceData);
+        }
     }
 
     @Override