+++ PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutClient.java
@@ -1,5 +1,7 @@
 package com.braintreepayments.api;
 
+import android.content.Context;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
 import androidx.fragment.app.Fragment;
@@ -30,26 +32,17 @@ public class PayPalNativeCheckoutClient {
     private PayPalNativeCheckoutListener listener;
 
     /**
-     * @param fragment        a {@link Fragment
-     * @param braintreeClient a {@link BraintreeClient}
-     * @deprecated see
-     * {@link PayPalNativeCheckoutClient#PayPalNativeCheckoutClient(BraintreeClient)}
-     * <p>
-     * Create a new instance of {@link PayPalNativeCheckoutClient} from within a Fragment using a
-     * {@link BraintreeClient}.
+     * Initializes a new {@link PayPalNativeCheckoutClient} instance
+     *
+     * @param context an Android Context
+     * @param authorization a Tokenization Key or Client Token used to authenticate
      */
-    @Deprecated
-    public PayPalNativeCheckoutClient(@NonNull Fragment fragment,
-                                      @NonNull BraintreeClient braintreeClient) {
-        this(braintreeClient, new PayPalNativeCheckoutInternalClient(braintreeClient));
+    public PayPalNativeCheckoutClient(@NonNull Context context, @NonNull String authorization) {
+        this(new BraintreeClient(context, authorization));
     }
 
-    /**
-     * Create a new instance of {@link PayPalNativeCheckoutClient} using a {@link BraintreeClient}.
-     *
-     * @param braintreeClient a {@link BraintreeClient}
-     */
-    public PayPalNativeCheckoutClient(@NonNull BraintreeClient braintreeClient) {
+    @VisibleForTesting
+     PayPalNativeCheckoutClient(@NonNull BraintreeClient braintreeClient) {
         this(braintreeClient, new PayPalNativeCheckoutInternalClient(braintreeClient));
     }