+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## unreleased
 
+* Move from Braintree to PayPal analytics service (FPTI)
 * Breaking Changes
   * All Modules
     * Bump `minSdkVersion` to API 23
     * Bump target Java version to Java 11
+    * Remove `BraintreeClient` public constructors
+    * Update payment method constructor parameters from `braintreeClient` to `context` and 
+      `authorization`
     * Remove `BraintreeSharedPreferencesException`
     * Ugrapde Kotlin version to 1.9.10
     * Upgrade to Android Gradle Plugin 8